% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criterion.R
\name{aic}
\alias{aic}
\title{Akaike Information Criterion (AIC)}
\usage{
aic(model)
}
\arguments{
\item{model}{R model object}
}
\value{
aic value
}
\description{
Akaike Information Criterion (AIC)
}
\references{
https://en.wikipedia.org/wiki/Akaike_information_criterion
}
