% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criterion.R
\name{bic}
\alias{bic}
\title{Bayesian Information Criterion}
\usage{
bic(model)
}
\arguments{
\item{model}{Base R model}

\item{X_train}{Training data

 A 2D matrix of (observations, features)}

\item{y_train}{Target class for training data}
}
\value{
bic value (float)
}
\description{
This is an implementation of the backward selection algorithm
that can be used to select best features in model.
}
