% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fit_and_score}
\alias{fit_and_score}
\title{Fit and Score}
\usage{
fit_and_score(S, feature, algorithm, X_train, y_train, X_val, y_val, criterion)
}
\arguments{
\item{S}{list of features as found in `forward()` and `backward()`}

\item{feature}{feature to add or drop (int)}

\item{algorithm}{direction of feature selection. One of: 'forward', 'backward'}

\item{X_train}{a 2D matrix of (observations, features).}

\item{y_train}{a 1D array of target classes for X_train.}

\item{X_val}{a 2D matrix of (observations, features).}

\item{y_val}{a 1D array of target classes for X_validate}

\item{criterion}{model selection criterion (string). One of:
'r-squared', 'aic', 'bic'}
}
\value{
score of the model as a float
}
\description{
fit and scores the linear regression model
}
\keyword{internal}
