% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fitter}
\alias{fitter}
\title{Fitter}
\usage{
fitter(X_input, y_input)
}
\arguments{
\item{X0}{a matrix of features}

\item{y0}{a response vector}
}
\value{
a fitted lm() model
}
\description{
Fits data to a linear regression model.
}
\keyword{internal}
