% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward.R
\name{forward}
\alias{forward}
\title{Forward Selection Algorithm.}
\usage{
forward(X_train, y_train, X_val, y_val, min_change = 0.5, n_features = NULL,
  criterion = "r-squared", verbose = TRUE)
}
\arguments{
\item{X_train}{Training data. Represented as a 2D matrix or dataframe of (observations, features).}

\item{y_train}{Target class for training data. Represented as a 1D vector of target classes for \code{X_train}.
If \code{y_train} is a character string AND \code{X} is a dataframe, it will be extracted from \code{X}.}

\item{X_val}{Validation data. Represented as a 2D matrix or dataframe of (observations, features).}

\item{y_val}{Target class for validation data. Represented as a 1D vector of target classes for \code{X_val}.
If \code{y_val} is a character string AND \code{X} is a dataframe, it will be extracted from \code{X}.}

\item{min_change}{The smallest change in criterion score to be considered significant.
Note: \code{n_features} must be NULL if this is numeric.}

\item{n_features}{The number of features to select, expressed either as a proportion (0,1)
or whole number with range (0,total_features). Note: \code{min_change} must be NULL if this is numeric.}

\item{criterion}{Model selection criterion to measure relative model quality. Can be one of:
\itemize{
 \item 'aic': use Akaike Information Criterion
 \item 'bic': use Akaike Information Criterion
 \item 'r-squared': use coefficient of determination
}}

\item{verbose}{If \code{TRUE}, print additional information as selection occurs}
}
\value{
A vector of indices that represent the best features of the model.
}
\description{
This is an implementation of the forward selection algorithm in which you
start with a null model and iteratively add the most useful features.
This function is built for the specific case of forward selection in
linear regression.
}
\examples{
X_train <- matrix(runif(50, 0, 50), ncol=5)
y_train <- runif(10, 0, 50)
X_val <- matrix(runif(50, 0, 20), ncol=5)
y_val <- runif(10, 0, 20)
forward(X_train, y_train, X_val, y_train, min_change=0.1, criterion="r-squared")
forward(X_train, y_train, X_val, y_train, n_features=0.1, criterion="aic")

}
