% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{input_checks}
\alias{input_checks}
\title{Input Checks}
\usage{
input_checks(n_features, min_change, criterion)
}
\arguments{
\item{n_features}{the number of features to select. Floats will be regarded as proportions
of the total that must lie on (0,1). `min_change` must be None for `n_features` to
operate.}

\item{min_change}{The smallest change to be considered significant.
`n_features` must be None for `min_change` to operate.}

\item{criterion}{model selection criterion:
* 'r-squared': use R-squared as the criterion
* 'aic': use Akaike Information Criterion
* 'bic': use Bayesian Information Criterion}
}
\description{
checks that:
(a) the only one of the the two inputs, `n_features` and `min_change`, are non-None
(b) the remaining element is numeric and is strictly greater than zero
}
\keyword{internal}
