% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{input_data_checks}
\alias{input_data_checks}
\title{Check Data Input}
\usage{
input_data_checks(X, y)
}
\arguments{
\item{X}{input for either \code{X_train} or \code{X_val}. Expected to be a 2D numeric matrix.}

\item{y}{input for either \code{y_train} or \code{y_val}. Expected to be a 1D numeric matrix.}
}
\keyword{internal}
