% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\name{mtcars_data}
\alias{mtcars_data}
\title{Generating example data with mtcars.}
\usage{
mtcars_data(seed = 99)
}
\arguments{
\item{seed}{random seed to use.
Defaults to 99.}
}
\value{
X_train, y_train, X_val, y_val (as a list of dataframes)
}
\description{
Generates test data using base R's mtcars dataset.
The response variable `y` is horsepower (`hp`), while the remaining variables
represent the predictive features `X`.
}
