% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_n_features}
\alias{parse_n_features}
\title{Parsing features}
\usage{
parse_n_features(n_features, total)
}
\arguments{
\item{n_features}{Numeric value passed to forward or backward selection.}

\item{total}{Total features in the data}
}
\value{
Number of features to select. If initial \code{n_features} lies on (0, total_features),
it will be returned 'as is'. If \code{n_feature} is passed in as a proportion with range (0,1), it will
be converted to an integer representing proportion with respect to the total number of features.
}
\description{
Parse \code{n_features} for forward and backward selection.
Handles two cases:
\itemize{
\item if \code{n_features} is an integer, ensure it lies on (0, total).
\item if \code{n_features} is a float, ensure it lies on (0, 1).
}
}
\keyword{internal}
