% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criterion.R
\name{aic}
\alias{aic}
\title{Akaike Information Criterion (AIC)}
\usage{
aic(model, X = NULL, y = NULL)
}
\arguments{
\item{model}{A base R model object (e.g., \code{lm()})}

\item{X}{Validation data as a 2D matrix of (observations, features).
If \code{NULL}, extract \code{X} from \code{model}.}

\item{y}{True labels as a 1D vector.
If \code{NULL}, extract \code{y} from \code{model}.}
}
\value{
AIC value gets returned as a float.
}
\description{
The Akaike Information Criterion's objective is to prevent model
 overfitting by adding a penalty term which penalizes more complex models.
 Its formal definition is:
 \deqn{-2*ln(L)+2*k}
 where L is the maximized value of the likelihood function.
 A smaller AIC value suggests that the model is a better fit for the data.
}
\examples{
model <- lm(hp~., data=mtcars)
aic(model)

}
\references{
https://en.wikipedia.org/wiki/Akaike_information_criterion
}
