% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criterion.R
\name{aic}
\alias{aic}
\title{Akaike Information Criterion (AIC)}
\usage{
aic(model)
}
\arguments{
\item{model}{R model object}
}
\value{
AIC value gets returned as float if sample size is sufficient. 
If n/k < 40 where n is the number of observations and k is the number of features, 
AICc gets returned to adjust for small sample size.
}
\description{
The Akaike Information Criterion's objective is to prevent model
 overfitting by adding a penalty term which penalizes more compelx models. 
 Its formal definition is:
 \deqn{-2*ln(L)+2*k}
 where L is the maximized value of the likelihood function. 
 A smaller AIC value suggests that the model is a better fit for the data.
}
\references{
https://en.wikipedia.org/wiki/Akaike_information_criterion
}
