% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backward.R
\name{backward}
\alias{backward}
\title{Backward Selection Algorithm.}
\usage{
backward(X_train, y_train, X_val, y_val, n_features = 0.5,
  min_change = NULL, criterion = "r-squared", verbose = TRUE)
}
\arguments{
\item{X_train}{Training data. Represented as a 2D matrix or dataframe of (observations, features).}

\item{y_train}{Target class for training data. Represented as a 1D vector of target classes for \code{X_train}.
If y_train is a character string AND X is a dataframe, it will be extracted from X.}

\item{X_val}{Validation data. Represented as a 2D matrix or dataframe of (observations, features).}

\item{y_val}{Target class for validation data. Represented as a 1D vector of target classes for \code{X_val}.
If y_val is a character string AND X is a dataframe, it will be extracted from X.}

\item{n_features}{The number of features to select, expressed either as a proportion (0,1)
or whole number with range (0,total_features)}

\item{min_change}{Smallest change in criterion score to be considered significant, as a numeric.}

\item{criterion}{Model selection criterion to measure relative model quality. Can be one of:
\itemize{
 \item 'aic': use Akaike Information Criterion
 \item 'bic': use Akaike Information Criterion
 \item 'r-squared': use coefficient of determination
}}

\item{verbose}{if \code{TRUE}, print additional information as selection occurs}
}
\value{
A vector of indices that represent the best features of the model.
}
\description{
This is an implementation of the backward selection algorithm
in which you start with a full model and iteratively remove the
least useful feature at each step. This function is built for the specific case
of backward selection in linear regression models.
}
