% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backward.R
\name{backward}
\alias{backward}
\title{Backward Selection Algorithm.}
\usage{
backward(X_train, y_train, X_val, y_val, n_features = 0.5,
  min_change = NULL, criterion = "r-squared", verbose = TRUE)
}
\arguments{
\item{X_train}{Training data

 A 2D matrix of (observations, features)}

\item{y_train}{Target class for training data

 A 1D array of target classes for X_train}

\item{X_val}{Validation data

 A 2D matrix of (observations, features)}

\item{y_val}{Target class for validation data

 A 1D array of target classes for X_val}

\item{n_features}{Maximum number of features to allow.}

\item{min_change}{Smallest change in criterion score to be considered significant.}

\item{criterion}{Model selection criterion
A criterion to measure relative model quality.
'aic': use Akaike Information Criterion
'bic': use Bayesian Information Criterion}

\item{verbose}{if True, print additional information as selection occurs}
}
\value{
A vector
}
\description{
This is an implementation of the backward selection algorithm
in which you start with a full model and iteratively remove the
least useful feature at each step. This function is built for the specific case
of feature selection in linear regression models.
}
\examples{
model = lm(Sepal.Length~Sepal.Width+Petal.Length+Petal.Width, data=iris)
bakcward(model, X_train, y_train, min_change=0.3, criterion='aic')
}
