% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fit_and_score}
\alias{fit_and_score}
\title{Fit and Score}
\usage{
fit_and_score(S, feature, algorithm, X_train, y_train, X_val, y_val, criterion)
}
\arguments{
\item{S}{A vector representing the list of selected features in `forward()` and `backward()`}

\item{feature}{Feature to add or drop, expressed as an integer.}

\item{algorithm}{Direction of feature selection. One of: 'forward', 'backward'.}

\item{X_train}{Training data. Represented as a 2D matrix of (observations, features).}

\item{y_train}{Target class for training data. Represented as a 1D vector of target classes for \code{X_train}.}

\item{X_val}{Validation data. Represented as a 2D matrix of (observations, features).}

\item{y_val}{Target classe for validation data. Represented as a 1D vector of target classes for \code{X_val}.}

\item{criterion}{Model selection criterion to measure relative model quality. Can be one of:
\itemize{
 \item 'aic': use Akaike Information Criterion
 \item 'bic': use Akaike Information Criterion
 \item 'r-squared': use coefficient of determination
}}
}
\value{
Score of the model as a float.
}
\description{
Fits a linear regression model to the data and
scores the relative quality of the model using r-squared, aic, or bic.
}
\keyword{internal}
