% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fitter}
\alias{fitter}
\title{Fitter}
\usage{
fitter(X_input, y_input)
}
\arguments{
\item{X_input}{A 2D matrix of features}

\item{y_input}{A 1D vector representing the response variable}
}
\value{
A fitted \code{lm()} model
}
\description{
Fits data to a linear regression model.
}
\keyword{internal}
