% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward.R
\name{forward}
\alias{forward}
\title{Forward Selection Algorithm.}
\usage{
forward(X_train, y_train, X_val, y_val, min_change = 0.5, n_features = NULL,
  criterion = "r-squared", verbose = TRUE)
}
\arguments{
\item{X_train}{Training data

 A 2D matrix of (observations, features)}

\item{y_train}{Target class for training data

 A 1D array of target classes for X_train}

\item{X_val}{Validation data

 A 2D matrix of (observations, features)}

\item{y_val}{Target class for validation data

 A 1D array of target classes for X_val}

\item{min_change}{Smallest change in criterion score to be considered significant.}

\item{criterion}{Model selection criterion
A criterion to measure relative model quality.
'aic': use Akaike Information Criterion
'bic': use Bayesian Information Criterion}

\item{verbose}{if True, print additional information as selection occurs}

\item{max_features}{Maximum number of features to allow.}
}
\value{
A vector
}
\description{
This is an implementation of the forward selection algorithm in which you
start with a null model and iteratively add the most useful features.
This function is built for the specific case of forward selection in
linear regression models.
}
\examples{
model = lm(Sepal.Length~Sepal.Width+Petal.Length+Petal.Width, data=iris)
forward(model, X_train, y_train, min_change=0.3, criterion='aic')

}
