% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward.R
\name{forward}
\alias{forward}
\title{Forward Selection Algorithm.}
\usage{
forward(X_train, y_train, X_val, y_val,
min_change=0.5,criterion='r-squared', verbose=TRUE)
}
\arguments{
\item{X_train}{Training data. Represented as a 2D matrix of (observations, features).}

\item{y_train}{Target class for training data. Represented as a 1D vector of target classes for \code{X_train}.}

\item{X_val}{Validation data. Represented as a 2D matrix of (observations, features).}

\item{y_val}{Target classe for validation data. Represented as a 1D vector of target classes for \code{X_val}.}

\item{min_change}{The smallest change in criterion score to be considered significant.}

\item{n_features}{The number of features to allow, expressed either as a proporition (0,1)
or whole number with range (0,total_features)}

\item{criterion}{Model selection criterion to measure relative model quality. Can be one of:
\itemize{
 \item 'aic': use Akaike Information Criterion
 \item 'bic': use Akaike Information Criterion
 \item 'r-squared': use coefficient of determination
}}

\item{verbose}{if \code{TRUE}, print additional information as selection occurs}
}
\value{
A vector of indices that represent the best features of the model.
}
\description{
This is an implementation of the forward selection algorithm in which you
start with a null model and iteratively add the most useful features.
This function is built for the specific case of forward selection in
linear regression.
}
