% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_n_features}
\alias{parse_n_features}
\title{Parsing features}
\usage{
parse_n_features(n_features, total)
}
\arguments{
\item{n_features}{numeric value passed to forward or backward selection}

\item{total}{total features in the data}
}
\value{
number of features to select. If `n_features` and it lies on (0, `total`),
it will be returned 'as is'. Whole numbers must lie of [0, total].
}
\description{
Parse `n_features` for forward and backward selection.
Handles two cases: (a) if `n_features` is an int, ensure it lies on (0, `total`),
and (b) if `n_features` is a float, ensure it lies on (0, 1).
}
\keyword{internal}
