% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criterion.R
\name{r_squared}
\alias{r_squared}
\title{R-squared}
\usage{
r_squared(fit_model, X, y)
}
\arguments{
\item{fit_model}{A fitted model}

\item{X}{Feature data}

\item{y}{True labels of the response variable}
}
\value{
The coefficient of determination
}
\description{
Calculates the coefficient of determination.
}
\references{
http://scikit-learn.org/stable/modules/model_evaluation.html#r2-score-the-coefficient-of-determination
}
